Pass nyxd client to members provider
