+++ ephemera/src/lib.rs
@@ -61,8 +61,8 @@ pub mod peer {
 /// Ephemera membership. How to find other nodes in the cluster.
 pub mod membership {
     pub use super::network::members::{
-        ConfigMembersProvider, DummyMembersProvider, HttpMembersProvider, JsonPeerInfo, PeerInfo,
-        PeerSetting, Result,
+        ConfigMembersProvider, DummyMembersProvider, JsonPeerInfo, PeerInfo, PeerSetting,
+        ProviderError, Result,
     };
 }
