+++ nym-api/Cargo.toml
@@ -68,6 +68,7 @@ zeroize = { workspace = true }
 actix-web = "4"
 array-bytes = "6.0.0"
 chrono = { version = "0.4.24", default-features = false, features = ["clock"] }
+futures-util = "0.3.25"
 serde_derive = "1.0.149"
 tempfile = "3.3.0"
 uuid = { version = "1.3.0", features = ["serde", "v4"] }
